+++ include/unigui/widgets/combobox.h
@@ -3,6 +3,7 @@
 #include <string>
 #include <vector>
 #include <functional>
+#include <imgui.h>
 
 namespace unigui {
 class ComboBox : public Widget {
@@ -17,9 +18,12 @@ public:
     void SetOnChange(std::function<void(int)> callback);
     void SetEditable(bool on);
     void SetSearchable(bool on);
+    void SetItemIcon(int index, ImTextureID textureID);
+    ImTextureID GetItemIcon(int index) const;
 private:
     std::string label_;
     std::vector<std::string> items_;
+    std::vector<ImTextureID> icons_;
     int selected_;
     bool editable_ = false;
     bool searchable_ = false;